// apis/open-weather/Academy/Models/Training.ts
export class Training {
    Id: number;
    Naam: string;
    Omschrijving: string;
    Voorkennis: string;
    Opzet: string;
    Leerdoel: string;
    HasExamen: boolean;
    MaxCapaciteit: number;
    TrainingCategorieId: number;
    DoelgroepId: number;
    HasAmortisatie: boolean;
    Eigenaar: number;
    TrainingMateriaal: string;
    Studielast: number;
    IsZelfstudieMogelijk: boolean;
    CreatedOn: Date;
}